chatbot_llm-main/chatbot_llm-main/requirements.txt
streamlit
# groq
google-generativeai
requests
openai
nltk
# streamlit.components.v1